+++ packages/commonwealth/client/scripts/views/components/Avatar/Avatar.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Avatar = () => {
+  return <div>Avatar</div>;
+};
