+++ packages/commonwealth/client/scripts/views/components/Avatar/index.ts
@@ -1,0 +1,4 @@
+import { AvatarUpload } from './AvatarUpload';
+import { Avatar } from './Avatar';
+
+export { AvatarUpload, Avatar };
