+++ packages/commonwealth/client/scripts/views/pages/create_community/chain_input_rows.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 // import { CommunityType } from '.';
 
 import app from 'state';
-import { AvatarUpload } from 'views/components/avatar_upload';
+import { AvatarUpload } from 'views/components/Avatar';
 import { InputRow } from 'views/components/metadata_rows';
 import type { DropdownItemType } from '../../components/component_kit/cw_dropdown';
 import { CWDropdown } from '../../components/component_kit/cw_dropdown';
