+++ packages/commonwealth/client/scripts/views/pages/manage_community/chain_metadata_rows.tsx
@@ -10,7 +10,7 @@ import { notifyError, notifySuccess } from 'controllers/app/notifications';
 import { InputRow, SelectRow, ToggleRow } from 'views/components/metadata_rows';
 import type ChainInfo from '../../../models/ChainInfo';
 import type RoleInfo from '../../../models/RoleInfo';
-import { AvatarUpload } from '../../components/avatar_upload';
+import { AvatarUpload } from '../../components/Avatar';
 
 import { CWButton } from '../../components/component_kit/cw_button';
 import { CWDropdown } from '../../components/component_kit/cw_dropdown';
