+++ packages/commonwealth/client/styles/components/Avatar/AvatarUpload.scss
@@ -1,4 +1,4 @@
-@import '../shared';
+@import '../../shared';
 
 .AvatarUpload {
   align-items: center;
